third_party/blink/web_tests/svg/dynamic-updates/SVGLinearGradientElement-svgdom-y2-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'y2' property of the SVGLinearGradientElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS linearGradientElement.y2.baseVal.valueAsString is "1000%"
PASS linearGradientElement.y2.baseVal.valueAsString is "0%"
PASS successfullyParsed is true

TEST COMPLETE
